Добавлены проверки для доступа на определённые операции

+++ source/api/views.py
@@ -19,9 +19,14 @@ def get_token_view(request, *args, **kwargs):
 
 class QuoteListApiView(APIView):
     def get(self, request, *args, **kwargs):
-        quotes = Quote.objects.all()
-        serializer = QuoteSerializer(quotes, many=True)
-        return Response(serializer.data)
+        if request.user.has_perm('view_quote'):
+            quotes = Quote.objects.all()
+            serializer = QuoteSerializer(quotes, many=True)
+            return Response(serializer.data)
+        else:
+            quotes = Quote.objects.filter(is_status=True)
+            serializer = QuoteSerializer(quotes, many=True)
+            return Response(serializer.data)
 
 class QuoteCreateApiView(APIView):
     def post(self, request, *args, **kwargs):
@@ -36,24 +41,35 @@ class QuoteCreateApiView(APIView):
 class QuoteDetailApiView(APIView):
     def get(self, request, *args, **kwargs):
         quote = Quote.objects.get(pk=kwargs['pk'])
-        serializer = QuoteSerializer(quote)
-        return Response(serializer.data)
+        if request.user.has_perm('view_quote'):
+            serializer = QuoteSerializer(quote)
+            return Response(serializer.data)
+        if quote.is_status == True:
+            serializer = QuoteSerializer(quote)
+            return Response(serializer.data)
+        else:
+            return Response({"err":"К сожалению у вас нет доступа для просмотра"}, status=HTTPStatus.BAD_REQUEST)
+
 
 class QuoteUpdateApiView(APIView):
     def put(self, request, *args, **kwargs):
-        quote = Quote.objects.get(pk=kwargs['pk'])
-        serializer = QuoteSerializer(data=request.data, instance=quote)
-        try:
-            serializer.is_valid()
-            serializer.save()
-            return Response(serializer.data, status=HTTPStatus.OK)
-        except ValidationError as err:
-            return Response(data=err.detail, status=HTTPStatus.BAD_REQUEST)
+        if request.user.has_perm('change_quote'):
+            quote = Quote.objects.get(pk=kwargs['pk'])
+            serializer = QuoteSerializer(data=request.data, instance=quote)
+            try:
+                serializer.is_valid()
+                serializer.save()
+                return Response(serializer.data, status=HTTPStatus.OK)
+            except ValidationError as err:
+                return Response(data=err.detail, status=HTTPStatus.BAD_REQUEST)
+        else:
+            return Response({"err": "К сожалению у вас нет доступа для редактирования"}, status=HTTPStatus.BAD_REQUEST)
 #
 class QuoteDeleteApiView(APIView):
     def delete(self, request, *args, **kwargs):
-        print(kwargs)
-        quote = get_object_or_404(Quote, pk=kwargs['pk'])
-        print('quote', quote)
-        quote.delete()
-        return Response({'pk' : kwargs['pk']}, status=204)
+        if request.user.has_perm('delete_quote'):
+            quote = get_object_or_404(Quote, pk=kwargs['pk'])
+            quote.delete()
+            return Response({'pk' : kwargs['pk']}, status=204)
+        else:
+            return Response({"err": "К сожалению у вас нет доступа для удаления"}, status=HTTPStatus.BAD_REQUEST)
